@@ -1,13 +1,12 @@
 [package]
 name = "feruca-benchmarks"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 feruca = { git = "https://github.com/theodore-s-beers/feruca" }
 icu = "1.5.0"
 icu_testdata = "1.5.0"
-once_cell = "1.20.2"
 
 [dev-dependencies]
 criterion = "0.5.1"
