@@ -1,4 +1,4 @@
-use criterion::{criterion_group, criterion_main, Criterion};
+use criterion::{Criterion, criterion_group, criterion_main};
 use icu::collator::{Collator, CollatorOptions};
 use icu::locid::locale;
 
